GETP-168 feat: 피플 조회 기능의 피플 좋아요 수 조회 구현

+++ src/main/java/es/princip/getp/domain/people/query/dto/people/PublicDetailPeopleResponse.java
@@ -8,8 +8,8 @@ public record PublicDetailPeopleResponse(
     String nickname,
     String profileImageUri,
     PeopleType peopleType,
-    Integer completedProjectsCount,
-    Integer likesCount,
+    long completedProjectsCount,
+    long likesCount,
     PublicDetailPeopleProfileResponse profile
 ) {
 }
